# src/main/resources/doc/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Role Resource","description":"This is role Referential for all endpoints"},{"name":"User Ref","description":"This is user Referential for all endpoints"},{"name":"Prospect User Resource","description":"This is prospect user referential for all endpoints"},{"name":"Save Prospect User","description":"This is the endpoint to save a new prospect user"}],"paths":{"/user/update":{"put":{"tags":["User Ref"],"summary":"Update user","operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"User updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"400":{"description":"Invalid user ID or user not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/role/update":{"put":{"tags":["Role Resource"],"summary":"Update role","operationId":"updateRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Role"}}},"required":true},"responses":{"400":{"description":"Invalid role name or role not found","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"Role updated successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/user/save":{"post":{"tags":["User Ref"],"summary":"Save new user","operationId":"saveUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"201":{"description":"User saved successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/role/save":{"post":{"tags":["Role Resource"],"summary":"Save new role","operationId":"saveRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Role"}}},"required":true},"responses":{"400":{"description":"Invalid role name or role already exists","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"object"}}}},"201":{"description":"Role created successfully","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/prospect-user/save":{"post":{"tags":["Prospect User Resource","Save Prospect User"],"operationId":"saveProspectUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProspectUserDTO"}}},"required":true},"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid prospect user or prospect user already exists","content":{"*/*":{"schema":{"type":"object"}}}},"201":{"description":"Prospect user created successfully","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/signup":{"post":{"tags":["Authentication referential"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"201":{"description":"User created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/auth/signin":{"post":{"tags":["Authentication referential"],"operationId":"signin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"User successfully logged in","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/user/partial-update":{"patch":{"tags":["User Ref"],"summary":"Partially update user","operationId":"partialUpdateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"User updated successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid user ID or user not found","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/user/find/{username}":{"get":{"tags":["User Ref"],"summary":"Get user by username","operationId":"getUser","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User fetched successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"400":{"description":"Invalid username or user not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/user/all":{"get":{"tags":["User Ref"],"summary":"Get all users","operationId":"getAllUsers","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}},"200":{"description":"Users fetched successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"/role/get/{name}":{"get":{"tags":["Role Resource"],"summary":"Get role by name","operationId":"getRole","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Invalid role name or role not found","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"Role fetched successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/role/all":{"get":{"tags":["Role Resource"],"summary":"Get all roles","operationId":"getAllRoles","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"Roles fetched successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Role"}}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Role"}}}}}}}},"/prospect-user/all":{"get":{"tags":["Prospect User Resource"],"description":"This is the endpoint to get all prospect users","operationId":"getAllProspectUsers","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"Prospect users retrieved successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/user/delete/{id}":{"delete":{"tags":["User Ref"],"summary":"Delete user by ID","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"User deleted successfully","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid user ID or user not found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/role/delete":{"delete":{"tags":["Role Resource"],"summary":"Delete role by id","operationId":"deleteRole","parameters":[],"responses":{"400":{"description":"Invalid role name or role not found","content":{"*/*":{"schema":{"type":"object"}}}},"204":{"description":"Role deleted successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"Authority":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Role":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"authorities":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Authority"}}}},"User":{"type":"object","properties":{"createdBy":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"activated":{"type":"boolean"},"imageUrl":{"type":"string"},"roles":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Role"}},"enabled":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"credentialsNonExpired":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"}}},"UserDTO":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string","description":"Login"},"firstName":{"type":"string","description":"First name"},"lastName":{"type":"string","description":"Last name"},"email":{"type":"string","description":"Email address"},"phone":{"type":"string","description":"Phone number"},"password":{"type":"string","description":"Password"},"imageUrl":{"type":"string","description":"User image"},"roleList":{"type":"array","description":"User roles","items":{"type":"string","description":"User roles"}}}},"AddressDTO":{"type":"object","properties":{"createdBy":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"city":{"type":"string"},"country":{"type":"string"},"street":{"type":"string"},"zipCode":{"type":"string"},"longitude":{"type":"number","format":"double"},"latitude":{"type":"number","format":"double"}}},"ProspectUserDTO":{"type":"object","properties":{"createdBy":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"cin":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"patentNumber":{"type":"string"},"phone":{"type":"string"},"username":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressDTO"}}},"AuthRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}}}}}